requirements.txt
streamlit==1.21.0
openai==0.27.0
youtube_transcript_api==0.5.0
pyairtable==1.4.0
bcrypt==4.0.1
PyYAML==6.0